+++ Sources/KeyboardKit/_Deprecated/RowItem.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "KeyboardRowItem")
+public typealias RowItem = KeyboardRowItem